Refactor image ordering and featuredImg logic
